app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
httpx==0.27.2
anthropic==0.34.2
faster-whisper==1.0.3
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-dotenv==1.0.1
fpdf2==2.7.9
